Add shared helper to apply role claims
